refactor + create /jobposts endpoint

+++ back-end/src/main/java/com/example/jobboard/repos/Candidate_AuthRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Candidate_AuthRepository extends JpaRepository<Candidate_Auth, Integer> {
+public interface Candidate_AuthRepository extends JpaRepository<Candidate_Auth, Long> {
 }
