+++ back-end/src/main/java/com/example/jobboard/repos/JobPostCategoryTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobPostCategoryTypeRepository extends JpaRepository<JobPostCategoryType, Integer> {
+public interface JobPostCategoryTypeRepository extends JpaRepository<JobPostCategoryType, Long> {
 }
